==> src/notebooks/01_json_scraping.ipynb <==
import json

file_path= "C:\\Users\\tothj\\Documents\\VS Code\\data\\json_scraping\\spotify_playlist.json"

with open(file_path, "r") as file:
    data = json.load(file)

    

data

data["contents"]["items"][0]["name"]



for track in data ["contents"]["items"]:
        for artist in track["artists"]:
            print(artist["name"])
        print()

for track in data ["contents"]["items"]:
    print(track["playCount"])
    print(track["name"])
    print()
