==> src/notebooks/02_web_scraping.ipynb <==
from bs4 import BeautifulSoup   
import pandas as pd 
import requests

url = "https://en.wikipedia.org/wiki/Community_areas_in_Chicago"
headers = {"User-Agent": "Mozilla/5.0"}
response = requests.get(url=url, headers=headers)
soup = BeautifulSoup(response.content, "html.parser")   
response

for table in soup.find_all("table"):
    print(table.get("class"))
    

table = soup.find("table", class_="wikitable sortable plainrowheaders mw-datatable")

table

for row in table.find_all("tr")[2:-1]:
    print(row)
    print("-" * 50)    


for row in table.find_all("tr")[2:-1]:
    cells = row.find_all("td")
    print(cells[0].get_text(strip=True))

for row in table.find_all("tr")[2:-1]:
    header_cells = row.find_all("th")
    print(header_cells[0].get_text(strip=True))
    print()

for row in table.find_all("tr")[2:-1]:
    header_cells = row.find_all("th")
    cells = row.find_all("td")

    print(cells[0].get_text(strip=True))
    print(header_cells[0].get_text(strip=True))
    print()
    

data = []

for row in table.find_all("tr")[2:-1]:
    header_cells = row.find_all("th")
    cells = row.find_all("td")

    area_code=cells[0].get_text(strip=True)
    community_area_name=header_cells[0].get_text(strip=True)
    
    data.append({
        "Area Code": area_code,
        "Community Area Name": community_area_name
    })


data

cummunity_areas_df = pd.DataFrame(data)
cummunity_areas_df.head()   

cummunity_areas_df.to_csv("community_areas_master.csv", index=False)
