==> src/notebooks/06_taxi_data_mapping.ipynb <==
from datetime import datetime   
from dateutil.relativedelta import relativedelta

import pandas as pd
pd.set_option("display.max_columns", 30) 
import requests

current_datetime = datetime.now() - relativedelta(months=2)
formatted_datetime = current_datetime.strftime("%Y-%m-%d")

url = ( 
    f"https://data.cityofchicago.org/resource/ajtu-isnz.json?"
    f"$where=trip_start_timestamp >= '{formatted_datetime}T00:00:00' " 
    f"AND trip_start_timestamp <= '{formatted_datetime}T23:59:59'"
    f"&$limit=30000"
)
response = requests.get(url)
data = response.json()

len(data)  # Number of taxi trips on February 16, 2024

taxi_trips = pd.DataFrame(data)
taxi_trips.head()

taxi_trips.info()

taxi_trips.describe()

taxi_trips[taxi_trips["fare"].isna()].sample(5)

#### Transformation : deal with NaN values

taxi_trips.drop(["pickup_census_tract", "dropoff_census_tract"], axis=1, inplace=True)

taxi_trips.drop(["pickup_centroid_location", "dropoff_centroid_location"], axis=1, inplace=True)

taxi_trips.dropna(inplace=True)

taxi_trips.info()

#### Transformation: renaming

taxi_trips.rename(columns={"pickup_community_area": "pickup_community_area_id",
                            "dropoff_community_area": "dropoff_community_area_id"}, inplace=True)

taxi_trips.columns

#### Transformation: create helper columns for weather data dimensions

taxi_trips['trip_start_timestamp'] = pd.to_datetime(taxi_trips['trip_start_timestamp'])

taxi_trips["datetime_for_weather"] = taxi_trips["trip_start_timestamp"].dt.floor("h")

taxi_trips["datetime_for_weather"].unique()

taxi_trips['trip_start_date'] = pd.to_datetime(taxi_trips['trip_start_timestamp']).dt.date

taxi_trips['trip_start_date'].sample(5)

#### Check: join taxi and weather data

current_datetime = datetime.now() - relativedelta(months=2)
formatted_datetime = current_datetime.strftime("%Y-%m-%d")

formatted_datetime



url = "https://archive-api.open-meteo.com/v1/era5"

params = {
    "latitude": 41.85,
    "longitude": -87.65,
    "start_date": formatted_datetime,
    "end_date": formatted_datetime,
    "hourly":"temperature_2m,wind_speed_10m,rain,precipitation",
}

response = requests.get(url, params=params)
data = response.json()

weather_data = {
    "datetime":data["hourly"]["time"],
    "temperature":data["hourly"]["temperature_2m"],
    "wind_speed":data["hourly"]["wind_speed_10m"],
    "rain": data["hourly"]["rain"],
    "precipitation":data["hourly"]["precipitation"],
    }


weather_df = pd.DataFrame(weather_data)
weather_df["datetime"] = pd.to_datetime(weather_df["datetime"])

weather_df.head()

taxi_trips_with_weather = taxi_trips.merge(weather_df, 
                                          left_on="datetime_for_weather", right_on="datetime")

taxi_trips_with_weather.head()

#### Transformation: data type conversions

taxi_trips.info()


taxi_trips_original = taxi_trips.copy()

data_types = {
    "trip_end_timestamp": "datetime64[ns]",
    "trip_seconds": "int32",
    "trip_miles": "float",
    "pickup_community_area_id": "int8",
    "dropoff_community_area_id": "int8",
    "fare": "float",
    "tips": "float",
    "tolls": "float",
    "extras": "float",
    "trip_total": "float",
    "pickup_centroid_latitude": "float",
    "pickup_centroid_longitude": "float",
    "dropoff_centroid_latitude": "float",
    "dropoff_centroid_longitude": "float",
}

taxi_trips = taxi_trips.astype(data_types)

taxi_trips.info()

taxi_trips.describe()

##### Memory usage

orig_mem = taxi_trips_original.memory_usage(deep=True).sum() 
opt_mem = taxi_trips.memory_usage(deep=True).sum()

reduction = (orig_mem - opt_mem) / orig_mem * 100

print(f"Original: {orig_mem:,} bytes")
print(f"Optimized: {opt_mem:,} bytes")
print(f"Reduction: {reduction:.2f}%")


#### Sanity Checks

taxi_trips[taxi_trips["trip_end_timestamp"] == taxi_trips["trip_end_timestamp"].max()]

taxi_trips[taxi_trips["trip_seconds"] == taxi_trips["trip_seconds"].max()]

taxi_trips[taxi_trips["fare"] == taxi_trips["fare"].max()]

#### Creating dimension tables for Company and Payment Type

##### Payment Type dimension

dim_payment_type = taxi_trips["payment_type"].drop_duplicates().reset_index(drop=True)

dim_payment_type = pd.DataFrame(
    {
    "payment_type_id": range(1, len(dim_payment_type) + 1),
    "payment_type": dim_payment_type
    })

dim_payment_type

##### Company dimension

dim_company = taxi_trips["company"].drop_duplicates().reset_index(drop=True)

dim_company = pd.DataFrame(
    {
    "company_id": range(1, len(dim_company) + 1),
    "company": dim_company
    })

dim_company

dim_company.tail()

##### Merging back to the Fact table

fact_taxi_trips = taxi_trips.merge(dim_payment_type, on = "payment_type")
fact_taxi_trips = fact_taxi_trips.merge(dim_company, on = "company")


fact_taxi_trips.sample(5)

fact_taxi_trips.drop(["payment_type", "company"], axis=1, inplace=True)

fact_taxi_trips.sample(5)

dim_payment_type.to_csv("dim_payment_type.csv", index=False)
dim_company.to_csv("dim_company.csv", index=False)

##### Create update logic ford Dimension tables

dummy_payment_type_data = [
    {"payment_type": "Credit Card"},
    {"payment_type": "X"},
    {"payment_type": "Y"},
    {"payment_type": "X"},
]

dummy_payment_type_data_df = pd.DataFrame(dummy_payment_type_data)


todays_payment_types = pd.DataFrame(dummy_payment_type_data_df["payment_type"].unique(), columns=["payment_type"])
new_payment_types = todays_payment_types[~todays_payment_types["payment_type"].isin(dim_payment_type["payment_type"])]

if not new_payment_types.empty:
    
    max_id = dim_payment_type["payment_type_id"].max() 
    new_payment_types["payment_type_id"] = range(max_id + 1, max_id + 1 + len(new_payment_types))
    dim_payment_type = pd.concat([dim_payment_type, new_payment_types], ignore_index=True)
    
dim_payment_type

#Testing the logic with different data

dummy_payment_type_data = [
    {"payment_type": "Credit Card"},
    {"payment_type": "X"},
    {"payment_type": "Z"},
    {"payment_type": "Z"},
]

dummy_payment_type_data_df = pd.DataFrame(dummy_payment_type_data)

todays_payment_types = pd.DataFrame(dummy_payment_type_data_df["payment_type"].unique(), columns=["payment_type"])
new_payment_types = todays_payment_types[~todays_payment_types["payment_type"].isin(dim_payment_type["payment_type"])]

if not new_payment_types.empty:
    
    max_id = dim_payment_type["payment_type_id"].max() 
    new_payment_types["payment_type_id"] = range(max_id + 1, max_id + 1 + len(new_payment_types))
    dim_payment_type = pd.concat([dim_payment_type, new_payment_types], ignore_index=True)
    
dim_payment_type

dim_company.tail()

dim_company = taxi_trips["company"].drop_duplicates().reset_index(drop=True)

dim_company = pd.DataFrame(
    {
    "company_id": range(1, len(dim_company) + 1),
    "company": dim_company
    })

dummy_company_data = [
    {"company": "Metro Jet Taxi A."},
    {"company": "X"},
    {"company": "Z"},
    {"company": "Z"},
]

dummy_company_data_df = pd.DataFrame(dummy_company_data)

todays_companies = pd.DataFrame(dummy_company_data_df["company"].unique(), columns=["company"])
new_companies = todays_companies[~todays_companies["company"].isin(dim_company["company"])]

if not new_companies.empty:
    
    max_id = dim_company["company_id"].max() 
    new_companies["company_id"] = range(max_id + 1, max_id + 1 + len(new_companies))
    dim_company = pd.concat([dim_company, new_companies], ignore_index=True)
    
dim_company.tail()
